+++ perl/FSM/Composition/Port.pm
@@ -1,0 +1,25 @@
+package FSM::Composition::Port;
+
+use v5.20;
+use strict;
+use warnings;
+use feature qw(signatures);
+no warnings 'experimental::signatures';
+
+sub new ($class, %args) {
+    return bless {
+        name => $args{name},
+        direction => $args{direction},
+        width => $args{width} // 1,
+        type => $args{type},
+        raw_token => $args{raw_token},
+    }, $class;
+}
+
+sub name ($self) { return $self->{name} }
+sub direction ($self) { return $self->{direction} }
+sub width ($self) { return $self->{width} }
+sub type ($self) { return $self->{type} }
+sub raw_token ($self) { return $self->{raw_token} }
+
+1;